wip: moved banner code to banner executor
